fix styling issues and add i18n stuff
