+++ webapp/channels/src/components/admin_console/access_control/editors/table_editor/values_editor.scss
@@ -3,31 +3,31 @@
     width: 100%;
     
     .select__multi-value {
+        display: flex;
         height: 24px;
+        align-items: center;
         padding: 0;
-        margin: 2px;
         border: none;
         border-radius: 4px;
+        margin: 2px;
         background-color: rgba(var(--center-channel-color-rgb), 0.08);
-        display: flex;
-        align-items: center;
     }
     
     .select__multi-value__label {
+        padding: 0 8px;
         color: var(--center-channel-color);
         font-size: 12px;
-        padding: 0 8px;
         line-height: 16px;
     }
     
     .select__multi-value__remove {
+        display: flex;
+        height: 100%;
+        align-items: center;
+        padding: 0 4px;
         border-radius: 0 4px 4px 0;
         color: rgba(var(--center-channel-color-rgb), 0.56);
         cursor: pointer;
-        padding: 0 4px;
-        height: 100%;
-        display: flex;
-        align-items: center;
         
         &:hover {
             background-color: rgba(var(--center-channel-color-rgb), 0.16);
