+++ webapp/channels/src/components/admin_console/access_control/jobs/access_control_sync_job_table.scss
@@ -35,8 +35,8 @@
     .job-table__access-control {
         .job-table__table {
             max-height: 400px;
-            border: none;
             padding: 0px;
+            border: none;
         }
 
         .table > thead > tr > th,
