+++ webapp/channels/src/components/admin_console/access_control/modals/policy_test/test_modal.scss
@@ -1,0 +1,4 @@
+#testResultsModalLabel {
+    padding: 32px 0;
+    text-align: center;
+}
