+++ webapp/channels/src/components/admin_console/access_control/modals/user_sync/synced_user_list.tsx
@@ -10,7 +10,6 @@ import type {UserProfile} from '@mattermost/types/users';
 import type {ActionResult} from 'mattermost-redux/types/actions';
 
 import {UserGroupsSVG} from 'components/common/svg_images_components/user_groups_svg';
-
 import SearchableUserList from 'components/searchable_user_list/searchable_user_list_container';
 
 import type {ActionFuncAsync} from 'types/store';
@@ -76,10 +75,7 @@ export const SyncedUserList = ({userIds, noResultsMessageId, noResultsDefaultMes
                 aria-label='No users found'
             >
 
-                <UserGroupsSVG className='empty-state-svg' />
-                    width={100}
-                    height={100}
-                />
+                <UserGroupsSVG className='empty-state-svg'/>
                 <h3 className='primary-message'>
                     <FormattedMessage
                         id={noResultsMessageId}
