+++ webapp/channels/src/i18n/en.json
@@ -258,7 +258,6 @@
   "admin.access_control.cel_help_modal.important_notes_title": "Important Notes",
   "admin.access_control.cel_help_modal.subheader": "With CEL you can define conditions to filter user attributes and control resource access.",
   "admin.access_control.cel_help_modal.title": "Common Expression Language (CEL)",
-  "admin.access_control.cel.learnMore": "Learn more about creating access expressions with examples.",
   "admin.access_control.cel.line_and_column_number": "L{lineNumber}:{columnNumber}",
   "admin.access_control.cel.validateSyntax": "Validate syntax",
   "admin.access_control.cel.validating": "Validating...",
@@ -294,6 +293,7 @@
   "admin.access_control.policy.edit_policy.delete_policy.subtitle.has_resources": "Remove all assigned resources (eg. Channels) to be able to delete this policy",
   "admin.access_control.policy.edit_policy.delete_policy.title": "Delete policy",
   "admin.access_control.policy.edit_policy.policyName": "Access control policy name:",
+  "admin.access_control.policy.edit_policy.policyName.placeholder": "Add a unique policy name",
   "admin.access_control.policy.edit_policy.switch_to_advanced": "Switch to Advanced Mode",
   "admin.access_control.policy.edit_policy.switch_to_simple": "Switch to Simple Mode",
   "admin.access_control.policy.edit_policy.title": "Edit Access Control Policy",
@@ -305,6 +305,8 @@
   "admin.access_control.policy.save_policy_confirmation_title": "Save access control policy ",
   "admin.access_control.table_editor.add_attribute": "Add attribute",
   "admin.access_control.table_editor.attribute": "Attribute",
+  "admin.access_control.table_editor.blank_state": "Select a user attribute and values to create a rule",
+  "admin.access_control.table_editor.learnMore": "Learn more about creating access expressions with examples.",
   "admin.access_control.table_editor.operator": "Operator",
   "admin.access_control.table_editor.operator.contains": "contains",
   "admin.access_control.table_editor.operator.ends_with": "ends with",
